refactor: `io` 패키지 생성
입출력과 관련 있는 객체들을 `io` 패키지로 이동
