+++ src/main/java/racingcar/io/Input.java
@@ -1,4 +1,4 @@
-package racingcar.input;
+package racingcar.io;
 
 import java.util.List;
