+++ src/main/java/racingcar/io/Output.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.io;
 
 public interface Output {
