+++ src/test/java/racingcar/io/KeyboardInputParserTest.java
@@ -1,4 +1,4 @@
-package racingcar.input;
+package racingcar.io;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
