+++ src/test/java/racingcar/io/KeyboardInputTest.java
@@ -1,4 +1,4 @@
-package racingcar.input;
+package racingcar.io;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
